% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-variant-effect-prediction.R
\name{add_variant_effect_prediction}
\alias{add_variant_effect_prediction}
\title{Add variant effect prediction}
\usage{
add_variant_effect_prediction(cds, vcf, genome)
}
\arguments{
\item{cds}{Coding sequence coordinates}

\item{vcf}{Data frame of variants from a variant call format file}

\item{genome}{A genome reference compatible with \link{get_genomic_sequence}}
}
\description{
Experimental, use with caution
}
\details{
\code{cds}, \code{vcf}, and \code{genome} should have matching chromosome naming
conventions.

\code{cds} must contain the following columns: \code{tx}, \code{chr}, \code{strand}, \code{start},
\code{end} and \code{exon_frame} with rows corresponding to exon coordinates.
\code{exon_frame} can be added via \link{add_exon_details}.

\code{vcf} must contain the following standard VCF columns:
\code{CHROM}, \code{POS}, \code{REF}, \code{ALT}.
}
