% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding-variants.R
\name{coding_variants}
\alias{coding_variants}
\title{Identify coding variants in a VCF file.}
\usage{
coding_variants(vcf, cds, genome, separate_alleles = TRUE)
}
\description{
Identify coding variants in a VCF file.
}
