% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sequence.R
\name{get_genomic_sequence}
\alias{get_genomic_sequence}
\alias{get_genomic_variant}
\alias{get_coding_sequence}
\alias{get_coding_variant}
\title{Get genomic sequences given ranges}
\usage{
get_genomic_sequence(chr, strand, start, end, genome)

get_genomic_variant(chr, strand, start, end, vcf, genome)

get_coding_sequence(chr, strand, start, end, cds, genome)

get_coding_variant(chr, strand, start, end, cds, vcf, genome)
}
\arguments{
\item{chr}{\code{[character]}

Chromosome names. Must match names returned by \code{names(genome)}.}

\item{strand}{\code{[character]}

Sequence strands (+|-).}

\item{start}{\code{[integer]}

Start coordinates of ranges.}

\item{end}{\code{[integer]}

End coordinates of ranges}

\item{genome}{\code{[BSgenome|DNAStringSet]}

A reference genome. See Details.}
}
\description{
Get genomic sequences given ranges
}
\details{
The reference genome can be either a \code{BSgenome} object from a
BSgenome reference package (see \link[BSgenome:BSgenome]{BSgenome::BSgenome}), or a \code{DNAStringSet}
object (see \link[Biostrings:DNAStringSet]{Biostrings::DNAStringSet}). \code{BSgenome} objects offer faster
sequence aquisition, but are limited to existing BSgenome packages
(see \link[BSgenome:available.genomes]{BSgenome::available.genomes}) whereas \code{DNAStringSet} objects can
be easily created from any standard FASTA file using
\link[Biostrings:readDNAStringSet]{Biostrings::readDNAStringSet}.
}
