% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-variant-effect.R
\name{predict_variant_effect}
\alias{predict_variant_effect}
\title{Add variant effect prediction}
\usage{
predict_variant_effect(cds, vcf, genome)
}
\arguments{
\item{cds}{Coding sequence coordinates}

\item{vcf}{Data frame of variants from a variant call format file}

\item{genome}{A genome reference compatible with \link{get_genomic_sequence}}
}
\description{
Experimental, use with caution
}
\details{
\code{cds}, \code{vcf}, and \code{genome} should have matching chromosome naming
conventions.

\code{cds} must contain the following columns (rows represent exon coordinates):
\code{tx}, \code{exon}, \code{chr}, \code{strand}, \code{start}, and \code{end}

\code{vcf} must contain the following standard VCF columns (rows represent variants):
\code{CHROM}, \code{POS}, \code{REF}, \code{ALT}.
}
\examples{
if (requireNamespace('BSgenome.Hsapiens.UCSC.hg38')) {
  library(tidyverse)
  library(mutagenesis)

  # Example files provided in this package
  cds_file <- system.file('extdata/CDS.csv', package = 'mutagenesis')
  vcf_file <- system.file('extdata/VCF.vcf', package = 'mutagenesis')

  # Coding sequence coordinates, variants, and a reference genome
  # are required to predict variant effects
  cds    <- read_csv(cds_file)
  vcf    <- read_vcf(vcf_file)
  genome <- BSgenome.Hsapiens.UCSC.hg38::Hsapiens
  vep    <- predict_variant_effect(cds, vcf, genome)

  # An example summary of effects
  vep \%>\%
    select(
      gene, ID:INFO, ref_cds, alt_cds, ref_aa, alt_aa, mutation_type,
      exon_boundary_dist, vcf_start, vcf_end, exon_start, exon_end
    ) \%>\%
    distinct() \%>\%
    count(ref_aa, mutation_type) \%>\%
    arrange(mutation_type, -n)
}

}
