% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read a VCF table}
\usage{
read_vcf(file, col_types = NULL, callback = NULL, ...)
}
\arguments{
\item{file}{\code{[character(1)]} Path to VCF file. Supported file extensions:
\code{.vcf}, \code{.vcf.gz}, \code{.vcf.bz2}. Compressed files are unzipped for the
duration of the R session via \link[R.utils:gunzip]{R.utils::gunzip}, or \link[R.utils:bunzip2]{R.utils::bunzip2}.}

\item{col_types}{\code{[col_spec|character(1)|NULL]} Either NULL, or a valid
column specification (for more details, see \link[readr:read_tsv]{readr::read_tsv}). Defaults
to only read the following expected VCF columns:\preformatted{- CHROM [character]
- POS [integer]
- ID [character]
- REF [character]
- ALT [character]
- QUAL [double]
- FILTER [character]
- INFO [character]
}}

\item{callback}{\code{[function(x, pos)|NULL]} A function with arguments \code{x} and
\code{pos} that takes a dataframe as \code{x} and returns a dataframe (\code{pos} is used
internally). This is usefull for filtering or summarizing VCF files that
are otherwise too big to read into memory. See \link[readr:DataFrameCallback]{readr::DataFrameCallback}
and \link[readr:read_tsv_chunked]{readr::read_tsv_chunked} for examples. Note that when using a
callback, the default value of \code{chunk_size} for \link[readr:read_tsv_chunked]{readr::read_tsv_chunked}
has been increased to \code{1e6}. Defaults to \code{NULL}.}

\item{...}{Arguments passed to \link[readr:read_tsv]{readr::read_tsv}, or
\link[readr:read_tsv_chunked]{readr::read_tsv_chunked} if \code{callback} is not \code{NULL}.}
}
\value{
An object with class \code{vcf} (inherits \code{data.frame} class). VCF header
information is stored as an attribute accessible via \code{attr(x, 'vcf')}.
}
\description{
Read a VCF table
}
\references{
\itemize{
\item \href{https://faculty.washington.edu/browning/beagle/intro-to-vcf.html}{Introduction to VCF}
}
}
