% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-complement.R
\docType{data}
\name{reverse_complement}
\alias{reverse_complement}
\alias{reverse}
\alias{complement}
\alias{DNA}
\alias{RNA}
\title{Reverse and/or complement strings}
\format{\strong{DNA} - A named \code{character} vector. Names and values
correspond to complementary DNA bases.

\strong{RNA} - A named \code{character} vector. Names and values
correspond to complementary RNA bases.}
\usage{
reverse_complement(x, from_to = DNA)

reverse(x)

complement(x, from_to = DNA)

DNA

RNA
}
\arguments{
\item{x}{\code{[character]}

Strings to reverse and/or complement.}

\item{from_to}{\code{[named:character]}

Names and values correspond to letters and their complement. Defaults
to \code{DNA}.}
}
\value{
Returns a character vector with the same length as \code{x}.
}
\description{
Given a character vector, returns the reverse, complement, or reverse &
complement of each string as a character vector.
}
\examples{
dna <- c('ATGATGC', 'AAAGGG')
reverse(dna)
complement(dna)
reverse_complement(dna)

rna <- c('AUGAUGC', 'AAAGGG')
reverse(rna)
complement(rna, RNA)
reverse_complement(rna, RNA)

}
\keyword{datasets}
